Yegor

+++ test/com/in6k/tests/first/ReapetablePrincepleTest.java
@@ -11,13 +11,19 @@ import java.nio.file.Paths;
 public class ReapetablePrincepleTest {
     @Test
     public void createFile() throws Exception {
-        Path path = Paths.get("/Users/alexander/testfile.txt");
+        Path path = Paths.get("/home/employee/testfile.txt");
         File file = path.toFile();
 
-        boolean created = file.createNewFile();
-        assertTrue("File must not exists before.", created);
+        if(file.exists()) {
+            boolean created = file.createNewFile();
+            assertTrue("File must not exists before.", created);
+        }
 
-        boolean deleted = file.delete();
-        assertTrue("File must be deleted.", deleted);
+        if(file.exists()) {
+            boolean deleted = file.delete();
+            assertTrue("File must be deleted.", deleted);
+        }
+
+
     }
 }
